+++ src/main/java/de/dealog/msg/service/MessageService.java
@@ -3,6 +3,7 @@ package de.dealog.msg.service;
 import de.dealog.msg.persistence.Message;
 import de.dealog.msg.persistence.MessageEntity;
 import de.dealog.msg.persistence.MessageRepository;
+import de.dealog.msg.persistence.MessageStatus;
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
 import io.quarkus.panache.common.Parameters;
 import io.quarkus.panache.common.Sort;
@@ -14,7 +15,6 @@ import org.geolatte.geom.Point;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import java.util.List;
-import java.util.Optional;
 
 /**
  *
@@ -33,19 +33,18 @@ public class MessageService {
      * @param size
      * @return
      */
-    public PagedList<? extends Message> list(final Point<G2D> point,
-                                             final int page, final int size) {
+    public PagedList<? extends Message> list(final Point<G2D> point, final int page, final int size) {
         log.debug("List messages for page {}, size {} and point '{}' ...", page, size, point);
         PanacheQuery<MessageEntity> messageQuery;
 
-        Sort sort = Sort.by("publishedAt", Sort.Direction.Descending);
-        String query = "";
-        Parameters parameters = null;
+        Sort sort = Sort.descending("publishedAt");
+        StringBuilder queryBuilder = new StringBuilder("status = :status");
+        Parameters parameters = Parameters.with("status", MessageStatus.Published);
         if (point != null) {
-            parameters = Parameters.with("point", point);
-            query = "within(:point, geocode) = true";
+            parameters.and("point", point);
+            queryBuilder.append(" within(:point, geocode) = true");
         }
-        messageQuery = messageRepository.find(query, sort, Optional.ofNullable(parameters).orElse(new Parameters()));
+        messageQuery = messageRepository.find(queryBuilder.toString(), sort, parameters);
 
         List<MessageEntity> list = messageQuery.page(page, size).list();
         long count = messageQuery.count();
@@ -63,15 +62,49 @@ public class MessageService {
     }
 
     /**
-     *
-     * @param message
+     * Create a new ...
+     * @param message {@link MessageEntity}
      */
     public void create(final Message message) {
         log.debug("Create message {}" , message);
         if (messageRepository.findByIdentifier(message.getIdentifier()) == null) {
+            message.setStatus(MessageStatus.Published);
             messageRepository.persistAndFlush((MessageEntity) message);
         } else {
             log.error("Found duplicate identifier {}", message.getIdentifier());
+        }
+    }
+
+    /**
+     * Message is updated
+     * @param message The message
+     */
+    public void update(final Message message) {
+        log.debug("Update message {}" , message);
+
+        MessageEntity byIdentifier = messageRepository.findByIdentifier(message.getIdentifier());
+        if (byIdentifier != null) {
+            byIdentifier.setHeadline(message.getHeadline());
+            byIdentifier.setDescription(message.getDescription());
+            byIdentifier.setGeocode(message.getGeocode());
+            messageRepository.persistAndFlush(byIdentifier);
+        } else {
+            log.error("Message for identifier {} not found", message.getIdentifier());
+        }
+    }
+
+    /**
+     * Message is superseded so update state to {@link MessageStatus#Superseded}
+     * @param message The message
+     */
+    public void supersede(final Message message) {
+        log.debug("Supersede message {}" , message);
+        MessageEntity byIdentifier = messageRepository.findByIdentifier(message.getIdentifier());
+        if (byIdentifier != null) {
+            byIdentifier.setStatus(MessageStatus.Superseded);
+            messageRepository.persistAndFlush(byIdentifier);
+        }else {
+            log.error("Message for identifier {} not found", message.getIdentifier());
         }
     }
 }