+++ src/test/java/de/dealog/msg/service/MessageServiceTest.java
@@ -2,7 +2,9 @@ package de.dealog.msg.service;
 
 import de.dealog.msg.TestUtils;
 import de.dealog.msg.persistence.Message;
+import de.dealog.msg.persistence.MessageEntity;
 import de.dealog.msg.persistence.MessageRepository;
+import de.dealog.msg.persistence.MessageStatus;
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
 import io.quarkus.panache.common.Parameters;
 import io.quarkus.panache.common.Sort;
@@ -10,12 +12,14 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.mockito.InjectMock;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
 import javax.inject.Inject;
 import java.util.Arrays;
 import java.util.List;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
@@ -23,10 +27,9 @@ import static org.mockito.Mockito.when;
 @QuarkusTest
 class MessageServiceTest {
 
-    public static final int PAGE_SIZE_3 = 3;
-
-    public static final long TOTAL_COUNT = Long.MAX_VALUE;
     public static final int PAGE_NUMBER_ONE = 1;
+    public static final int PAGE_SIZE_3 = 3;
+    public static final long TOTAL_COUNT = Long.MAX_VALUE;
 
     @Inject
     MessageService messageService;
@@ -57,5 +60,35 @@ class MessageServiceTest {
         Assertions.assertEquals(messages.size(), list.getPageSize());
         Assertions.assertEquals(TOTAL_COUNT, list.getCount());
         Assertions.assertEquals(pageCount, list.getPageCount());
+    }
+
+    @Test
+    void update() {
+        MessageEntity persisted = TestUtils.buildMessage("1", "This is the headline", "This is the description");
+        Message updated = TestUtils.buildMessage("1", "This is a new headline", "This is a new description");
+        when(messageRepository.findByIdentifier("1")).thenReturn(persisted);
+        messageService.update(updated);
+
+        ArgumentCaptor<MessageEntity> messageCaptor = ArgumentCaptor.forClass(MessageEntity.class);
+        Mockito.verify(messageRepository, Mockito.times(1)).persistAndFlush(messageCaptor.capture());
+        MessageEntity capturedMessage = messageCaptor.getValue();
+        assertEquals(updated.getIdentifier(), capturedMessage.getIdentifier());
+        assertEquals(updated.getHeadline(), capturedMessage.getHeadline());
+        assertEquals(updated.getDescription(), capturedMessage.getDescription());
+    }
+
+    @Test
+    void supersede() {
+        MessageEntity persisted = TestUtils.buildMessage("1", "This is the headline", "This is the description");
+        persisted.setStatus(MessageStatus.Published);
+        when(messageRepository.findByIdentifier("1")).thenReturn(persisted);
+
+        Message superseded = TestUtils.buildMessage("1", "This is the headline", "This is the description");
+        messageService.supersede(superseded);
+
+        ArgumentCaptor<MessageEntity> messageCaptor = ArgumentCaptor.forClass(MessageEntity.class);
+        Mockito.verify(messageRepository, Mockito.times(1)).persistAndFlush(messageCaptor.capture());
+        MessageEntity capturedMessage = messageCaptor.getValue();
+        assertEquals(MessageStatus.Superseded, capturedMessage.getStatus());
     }
 }